Multiple small fixes
Should write more here but currently pretty lazy.
Mostly md3 importing fixes from the entity tab...
